+++ src/main/java/com/elmakers/mine/bukkit/magic/MagicController.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.FilenameFilter;
 import java.io.InputStream;
 import java.lang.ref.WeakReference;
+import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.net.URL;
@@ -583,8 +584,32 @@ public class MagicController implements Listener, MageController
         crafting = new CraftingController(this);
 		enchanting = new EnchantingController(this);
 		anvil = new AnvilController(this);
-		load();
-		
+
+        // Check for EffectLib.
+        Object effectLib = plugin.getServer().getPluginManager().getPlugin("EffectLib");
+        hasEffectLib = false;
+        if (effectLib != null) {
+            // Unfortunately, EffectLib is not Mavenized so I can't
+            // use its API without breaking my build.
+            try {
+                Class<?> effectManagerClass = Class.forName("de.slikey.effectlib.EffectManager");
+                Constructor managerConstructor = effectManagerClass.getConstructor(effectLib.getClass());
+                Object effectManager = managerConstructor.newInstance(effectLib);
+                hasEffectLib = EffectPlayer.setEffectManager(effectManager);
+            } catch (Exception ex) {
+                getLogger().warning("Found EffectLib, but encountered an error integrating: " + ex.getMessage());
+                hasEffectLib = false;
+            }
+        }
+
+        if (hasEffectLib) {
+            getLogger().info("EffectLib found, will be used for effects");
+        } else {
+            getLogger().info("EffectLib not found, install it for better effects");
+        }
+
+        load();
+
 		// Try to link to Essentials:
 		Object essentials = plugin.getServer().getPluginManager().getPlugin("Essentials");
 		hasEssentials = essentials != null;
@@ -780,6 +805,9 @@ public class MagicController implements Listener, MageController
 			   
 			    if (metricsLevel > 1) {
 			    	Graph integrationGraph = metrics.createGraph("Plugin Integration");
+                    integrationGraph.addPlotter(new Metrics.Plotter("EffectLib") {
+                        @Override public int getValue() { return controller.hasEffectLib ? 1 : 0; }
+                    });
 			    	integrationGraph.addPlotter(new Metrics.Plotter("Essentials") {						
 						@Override public int getValue() { return controller.hasEssentials ? 1 : 0; }
 					});
@@ -2175,8 +2203,8 @@ public class MagicController implements Listener, MageController
 		} else if (mage.isNewPlayer() && welcomeWand.length() > 0) {
             mage.clearNewPlayer();
 			wand = Wand.createWand(this, welcomeWand);
-            wand.takeOwnership(player, false, false);
 			if (wand != null) {
+                wand.takeOwnership(player, false, false);
 				giveItemToPlayer(player, wand.getItem());
 				getLogger().info("Gave welcome wand " + wand.getName() + " to " + player.getName());
 			} else {
@@ -3263,7 +3291,8 @@ public class MagicController implements Listener, MageController
 	 private Metrics							 metrics						= null;
 	 private boolean							 hasDynmap						= false;
 	 private boolean							 hasEssentials					= false;
-	 private boolean							 hasCommandBook					= false;
+     private boolean							 hasEffectLib					= false;
+     private boolean							 hasCommandBook					= false;
 	 private boolean							 hasWorldEdit					= false;
 	 
 	 // Sub-Controllers